refactor: 헥사고날 구조 제거
