+++ STORIX-Domain/src/main/java/com/storix/domain/domains/topicroom/adaptor/TopicRoomPersistenceAdapter.java
@@ -22,9 +22,7 @@ import org.springframework.data.domain.Slice;
 import org.springframework.stereotype.Component;
 
 import java.time.LocalDateTime;
-import java.util.Collections;
 import java.util.List;
-import java.util.Set;
 
 @Component
 @RequiredArgsConstructor
@@ -115,22 +113,6 @@ public class TopicRoomPersistenceAdapter implements LoadTopicRoomPort, RecordTop
     }
 
     @Override
-    public List<TopicRoom> loadTop5PopularRooms() {
-        return topicRoomRepository.findTop5ByOrderByPopularityScoreDescLastChatTimeDesc();
-    }
-
-    @Override
-    public Set<Long> loadJoinedRoomIds(Long userId, List<Long> roomIds) {
-
-        // 빈 리스트일 경우 -> 빈 Set 반환
-        if (roomIds == null || roomIds.isEmpty()) {
-            return Collections.emptySet();
-        }
-
-        return topicRoomUserRepository.findJoinedRoomIdsByUserIdAndRoomIds(userId, roomIds);
-    }
-
-    @Override
     public List<Long> findAllJoinedRoomIdsByUserId(Long userId) {
         return topicRoomUserRepository.findAllJoinedRoomIdsByUserId(userId);
     }
@@ -146,11 +128,6 @@ public class TopicRoomPersistenceAdapter implements LoadTopicRoomPort, RecordTop
     @Override
     public boolean existsById(Long roomId) {
         return topicRoomRepository.existsById(roomId);
-    }
-
-    @Override
-    public boolean existsByUserIdAndRoomId(Long userId, Long roomId) {
-        return topicRoomUserRepository.existsByUserIdAndTopicRoomId(userId, roomId);
     }
 
     @Override
