+++ STORIX-Domain/src/main/java/com/storix/domain/domains/topicroom/application/port/LoadTopicRoomPort.java
@@ -28,7 +28,5 @@ public interface LoadTopicRoomPort {
 
     boolean existsById(Long roomId);
 
-    List<TopicRoom> loadTop5PopularRooms();
-
     List<TopicRoom> findAllActiveRooms();
 }
