+++ STORIX-Domain/src/main/java/com/storix/domain/domains/topicroom/service/TopicRoomService.java
@@ -6,6 +6,7 @@ import com.storix.domain.domains.search.dto.PlusSearchResponseWrapperDto;
 import com.storix.domain.domains.search.dto.SearchResponseWrapperDto;
 import com.storix.domain.domains.search.dto.TrendingItem;
 import com.storix.domain.domains.search.service.SearchHistoryService;
+import com.storix.domain.domains.topicroom.adaptor.TopicRoomAdaptor;
 import com.storix.domain.domains.topicroom.application.port.LoadTopicRoomUserPort;
 import com.storix.domain.domains.topicroom.application.port.LoadTopicRoomPort;
 import com.storix.domain.domains.topicroom.application.port.RecordTopicRoomPort;
@@ -22,6 +23,7 @@ import com.storix.domain.domains.topicroom.exception.*;
 import com.storix.domain.domains.user.adaptor.UserAdaptor;
 import com.storix.domain.domains.user.application.port.LoadUserPort;
 import com.storix.domain.domains.user.domain.User;
+import com.storix.domain.domains.works.adaptor.WorksAdaptor;
 import com.storix.domain.domains.works.application.port.LoadWorksPort;
 import com.storix.domain.domains.works.domain.Genre;
 import com.storix.domain.domains.works.domain.Works;
@@ -56,6 +58,8 @@ public class TopicRoomService implements TopicRoomUseCase {
     private final LoadTopicRoomUserPort loadTopicRoomMemberPort;
     private final GenreScorePublisher genreScorePublisher;
     private final UserAdaptor userAdaptor;
+    private final TopicRoomAdaptor topicRoomAdaptor;
+    private final WorksAdaptor worksAdaptor;
 
     @Override
     public Slice<TopicRoomResponseDto> getMyJoinedRooms(Long userId, Pageable pageable) {
@@ -69,7 +73,7 @@ public class TopicRoomService implements TopicRoomUseCase {
                 .toList();
 
         // works 정보를 한 번에 조회하여 Map으로 변환
-        Map<Long, TopicRoomWorksInfo> worksMap = loadWorksPort.loadWorksMapByIds(worksIds);
+        Map<Long, TopicRoomWorksInfo> worksMap = worksAdaptor.loadWorksMapByIds(worksIds);
 
         return participations.map(participation -> {
             TopicRoom room = participation.getTopicRoom();
@@ -107,7 +111,7 @@ public class TopicRoomService implements TopicRoomUseCase {
     @Override
     public List<TopicRoomPreviewResponseDto> getPopularRooms(Long userId) {
         // 1. 상위 5개 토픽룸 가져오기
-        List<TopicRoom> rooms = loadTopicRoomPort.loadTop5PopularRooms();
+        List<TopicRoom> rooms = topicRoomAdaptor.loadTop5PopularRooms();
         if (rooms.isEmpty()) return Collections.emptyList();
 
         List<Long> roomIds = rooms.stream().map(TopicRoom::getId).toList();
@@ -118,12 +122,12 @@ public class TopicRoomService implements TopicRoomUseCase {
                 .distinct()
                 .toList();
 
-        Map<Long, TopicRoomWorksInfo> worksMap = loadWorksPort.loadWorksMapByIds(worksIds);
+        Map<Long, TopicRoomWorksInfo> worksMap = worksAdaptor.loadWorksMapByIds(worksIds);
         Map<Long, String> nicknameMap = userAdaptor.findNicknameMapByUserIds(senderIds);
 
         // 포트를 통해 Set<Long> 형태의 가입된 방 ID 목록 수신
         Set<Long> joinedRoomIds = (userId != null)
-                ? loadTopicRoomMemberPort.loadJoinedRoomIds(userId, roomIds)
+                ? topicRoomAdaptor.loadJoinedRoomIds(userId, roomIds)
                 : Collections.emptySet();
 
         return rooms.stream()
