+++ STORIX-Domain/src/main/java/com/storix/domain/domains/works/adaptor/WorksAdaptor.java
@@ -1,8 +1,15 @@
 package com.storix.domain.domains.works.adaptor;
 
+import com.storix.domain.domains.works.dto.TopicRoomWorksInfo;
 import com.storix.domain.domains.works.repository.WorksRepository;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Component;
+
+import java.util.Collections;
+import java.util.List;
+import java.util.Map;
+import java.util.function.Function;
+import java.util.stream.Collectors;
 
 @Component
 @RequiredArgsConstructor
@@ -12,5 +19,17 @@ public class WorksAdaptor {
 
     public long countAllWorks() {
         return worksRepository.count();
+    }
+
+    public Map<Long, TopicRoomWorksInfo> loadWorksMapByIds(List<Long> worksIds) {
+
+        if (worksIds == null || worksIds.isEmpty()) {
+            return Collections.emptyMap();
+        }
+
+        List<TopicRoomWorksInfo> infos = worksRepository.findSimpleInfoByIdIn(worksIds);
+
+        return infos.stream()
+                .collect(Collectors.toMap(TopicRoomWorksInfo::id, Function.identity()));
     }
 }
