+++ STORIX-Domain/src/main/java/com/storix/domain/domains/works/application/port/LoadWorksPort.java
@@ -52,8 +52,6 @@ public interface LoadWorksPort {
     // 피드 관심 작품 리스트 용
     Map<Long, SlicedWorksInfo> findAllSlicedWorksInfoByWorksIds(List<Long> worksIds);
 
-    Map<Long, TopicRoomWorksInfo> loadWorksMapByIds(List<Long> worksIds);
-
     // 랜덤 작품 조회 (없으면 빈 리스트)
     List<Works> findRandomWorksExcluding(List<Long> excludedIds, int limit);
